refactor: Reduce zb binary size
